corrected mistakes in creare-element.js
