+++ index.html
@@ -163,7 +163,7 @@
   <!-- Шаблон изображения случайного пользователя -->
   <template id="picture">
     <a href="#" class="picture">
-      <img class="picture__img" src="./photos/1.jpg" width="182" height="182" alt="Случайная фотография" />
+      <img class="picture__img" src="" width="182" height="182" alt="Случайная фотография" />
       <p class="picture__info">
         <span class="picture__comments"></span>
         <span class="picture__likes"></span>
